sandbox/requirements.txt
Django==1.6
django-oscar==0.7.3
django-debug-toolbar==1.2.1
django-extensions==1.4.0
#django-oscar-amazon-payments==0.1
